# openapi.yaml
...
openapi: 3.0.3
info:
  title: blunderDB engine API
  version: "1"
  description: >-
    POST /v1/<family>.<method> over HTTP + JSON (NDJSON for a streaming
    list). SECURITY: this daemon performs NO authentication of its own — it
    trusts the X-Tenant-ID request header verbatim and MUST run behind an
    authenticating reverse-proxy (see ADR-0005 and doc/source/mode_headless.rst).
    X-Tenant-ID is the tenant's positive decimal integer; a name is refused
    with 400 invalid, never mapped to a tenant by this daemon. Idempotency:
    positions.save (and every other positions.* write) is naturally
    idempotent through its Zobrist content hash — saving the same position
    twice returns the same row, never a duplicate. collections.create,
    tournaments.create and anki.reviewCard have no such natural key (two
    calls are two distinct effects) and instead accept an optional
    Idempotency-Key request header (marked "x-idempotency-key: true" below):
    a retried call carrying the same key replays the first attempt's result
    instead of repeating its effect. Every other route neither needs nor
    accepts that header.
servers:
  - url: /v1
paths:
...
  /v1/directions.lastDecision:
    post:
      operationId: directions.lastDecision
      tags: [directions]
      x-kind: json
      parameters:
        - name: If-None-Match
          in: header
          required: false
          description: An ETag of an earlier answer; 304 while it still holds.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/directionReq"
      responses:
        "304":
          description: Not Modified — the If-None-Match tag still holds; no body.
        "200":
          description: OK
          headers:
            ETag:
              description: Weak tag of this answer; send it back in If-None-Match.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/service_LastDecision"
                nullable: true
...
  /v1/directions.standings:
    post:
      operationId: directions.standings
      tags: [directions]
      x-kind: json
      parameters:
        - name: If-None-Match
          in: header
          required: false
          description: An ETag of an earlier answer; 304 while it still holds.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/directionReq"
      responses:
        "304":
          description: Not Modified — the If-None-Match tag still holds; no body.
        "200":
          description: OK
          headers:
            ETag:
              description: Weak tag of this answer; send it back in If-None-Match.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/service_StandingsView"
                nullable: true
...
  /v1/tournaments.matches:
    post:
      operationId: tournaments.matches
      tags: [tournaments]
      x-kind: stream
      x-streaming: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tournamentIDReq"
      responses:
        "200":
          description: OK
          content:
            application/x-ndjson:
              schema:
                description: One JSON object per line (NDJSON), each shaped as below.
                $ref: "#/components/schemas/domain_Match"
                nullable: true
...
components:
  schemas:
...
    directionReq:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int64
...
    domain_Match:
      type: object
      description: "Opaque: see the Go type domain.Match (defined outside internal/server, not expanded here)."
...
    service_LastDecision:
      type: object
      description: "Opaque: see the Go type service.LastDecision (defined outside internal/server, not expanded here)."
...
    service_StandingsView:
      type: object
      description: "Opaque: see the Go type service.StandingsView (defined outside internal/server, not expanded here)."
...
    tournamentIDReq:
      type: object
      properties:
        tournamentId:
          type: integer
          format: int64
...